Doc: Remind people to install Yeoman :)
By calling out the fact that one needs to have Node and npm installed for any of this to work, the README implies those are the *only* prerequisites. Obviously you need Yeoman as well, but I can see how someone who's not experienced with Yeoman would assume it's just part of Node/npm or even a built-in command like `ls`.